fix: Journey now namespaced
