+++ app/graphql/types/tfl/entities/journey_planner/itinerary_result_type.rb
@@ -6,7 +6,7 @@ module Types
           field :stopMessages, [String, null: true], null: true
           field :recommendedMaxAgeMinutes, Integer, null: true
           field :lines, [Types::Tfl::Entities::LineType, null: true], null: true
-          field :journeys, [Types::JourneyType, null: true], null: true
+          field :journeys, [Types::Tfl::Entities::JourneyPlanner::JourneyType, null: true], null: true
           field :journeyVector, Types::Tfl::Entities::JourneyPlanner::JourneyVectorType, null: true
           field :searchCriteria, Types::SearchCriteriaType, null: true
           field :cycleHireDockingStationData, Types::CycleHireDockingStationDataType, null: true
